Time Slider: Move ThrottleAwaiter to shared location, fix whitespace and copyright header
